refactor: 模块化重构和打包配置修复
- feat: 完全重构为模块化架构，将功能拆分为独立模块
  - src/constants.js - 常量定义模块
  - src/utils.js - 工具函数模块
  - src/themes.js - 主题配置模块
  - src/config.js - 配置管理模块
  - src/commands.js - 命令处理模块
  - src/config-validator.js - 配置验证和监听模块

- fix: 修复打包配置问题，移除 .vscodeignore 中的 src/** 规则
  - 解决了打包时 src 文件夹被忽略的问题，确保所有模块正确打包

- fix: 修复命令注册问题，修正 extension.js 中的模块导入路径
  - 解决了 autoConfigureRainbowCursor 函数导入路径错误的问题

- docs: 更新 README.md 和 README.en.md
  - 更新版本号至 3.5.1
  - 添加模块化架构说明
  - 更新更新日志，详细记录本次重构的改进

- build: 更新 package.json 版本号至 3.5.1
