+++ package.json
@@ -2,7 +2,7 @@
   "name": "woodfish-theme",
   "displayName": "Woodfish Theme",
   "description": "🌈 Woodfish 彩虹主题 - 回档到稳定版本，具有发光效果、渐变色彩和动画的VSCode主题扩展",
-  "version": "3.4.0",
+  "version": "3.5.1",
   "publisher": "zhongjun",
   "author": {
     "name": "Woodfish"
